+++ app/src/main/java/com/soundcloud/android/model/PlayableProperty.java
@@ -6,7 +6,7 @@ import java.util.Date;
 
 public class PlayableProperty extends EntityProperty {
     public static final Property<String> TITLE = Property.of(PlayableProperty.class, String.class);
-    public static final Property<Integer> DURATION = Property.of(PlayableProperty.class, Integer.class);
+    public static final Property<Long> DURATION = Property.of(PlayableProperty.class, Long.class);
     public static final Property<Urn> CREATOR_URN = Property.of(PlayableProperty.class, Urn.class);
     public static final Property<String> CREATOR_NAME = Property.of(PlayableProperty.class, String.class);
     public static final Property<String> REPOSTER = Property.of(PlayableProperty.class, String.class);
